added initramfs-tools scripts for udev
these are from udev_164-3_amd64.deb

this'll mean that building the initrd will depend on udev being installed on
the host system, and that the initrd build process will refer to and copy over
the udev configuration from the host system.  If we build on a stock debian
install it should still be generic.