updated runeditor script with commented line to optionally run vim
so that the user can choose which editor to use before building the usb key
image.

this needs to be set before the image is built

One thing I would like to change about the vim version is that it is not
restricted to the /mnt/text/ directory on the usb key.  That means the
user can access and change the system files.  That might be a fun game,
but it's potentially distracting.